New Drag System
There is still a slight issue, such that you can not drag the Task Screen section up or down via holding a task card. You can move it using the space in between though.
